Module main::buscaminas_service

Service principal, ya que es el que contiene la logica mas importante del buscaminas. Se encuentra modularizado en las distintas funcionalidades requeridas para lograr develar las minas.

Structs

Buscaminas 🔒

Struct de Buscaminas Contiene todas las estructuras necesarias para facilitar la implementación

Constants

BOMBA_STRING

BOMBA_U8

ENTER_STRING

NUM_CERO_U8

SIN_BOMBAS_STRING

Functions

agregar_recuento_de_minas 🔒

Función que dado un Buscaminas, agrega a cada una de las posiciones de su respectivo mapa la cantidad correcta de minas adyacentes a dicha posicion, representado como u8. En caso de no poseer minas adyacentes deja la posicion vacia. Retorna el Buscaminas.

calcular_minas_adyacentes 🔒

Funcion que dada una posicion determinada en el mapa de Buscaminas, calcula la cantidad de minas adyacentes a esa posicion y la retorna.

crear_mapa_input 🔒

Funcion que a partir de una tira de bytes, ubica cada uno de ellos en su respectiva posicion correspondiente al mapa de Buscaminas. Retorna un Buscaminas.

descubrir_minas

Función principal. Dentro se encuentran modularizados los distintos llamados segun las funcionalidades requeridas para completar el descubrimiento de minas. Recibe el mapa input y retorna el mapa con el conteo de las minas adyacentes, ambos en formato String.

pasar_mapa_a_string 🔒

Funcion que dado un Buscaminas, obtiene y pasa su mapa de u8 a String y lo retorna.

quitar_enters 🔒

Funcion que suprime los enters de un string y lo retorna.