Module main::calculadora_service

Este modulo tiene como objetivo centralizar las operaciones y calculos realizados en base a las filas y columnas de la matriz del Buscaminas.

Todas las funciones de calculo de limites calculan el respectivo rango sobre el cual se podra iterar alrededor de una determinada posicion

Constants

ENTER_U8 🔒

Functions

calcular_col_limite_der

calcular_col_limite_izq

calcular_fil_limite_inf

calcular_fil_limite_sup

cant_columnas

Getter de la cantidad de columnas del Buscaminas. Al igual que el getter de las filas, también recibe la tira de bytes con los enters incluidos, razon por la cual al finalizar se le resta 1. Ejemplo de una matriz 2x3 de bombas:

cant_filas

Getter de la cantidad de filas del Buscaminas. La tira de bytes recibida contiene un enter al finalizar cada fila. Ejemplo para una matriz 2x2 de bombas: