Module main::catalog

Defines lab parameters LabParams and LabDef, the parameters for running the test diagram from the Adapton Lab README. This module of Adapton Lab extends Adapton's existing reflection (see adapton::engine::reflect) with further reflections. In particular, we produce HTML output structure, for human user interaction and consumption. Generically implements the test diagram in the Adapton Lab README. Provides concrete instances of the test diagram from the Adapton Lab README. Provides concrete instances of the test diagram from the Adapton Lab README.

Modules

hammer_s17_hw0

hammer_s17_hw1

oopsla2015_sec2
Program list_map running on a small, changing input list. It is simpler than the version from the Adapton collections library.

Structs

EagerFilter

EagerMap

EagerMap2

EagerMergesort1

EagerMergesort2

EagerMergesort3

EditorOopsla2015Sec2
This list editor mimics the editor in the example from Incremental Computation with Names (2015), Section 2 (Figs 1 and 2).

ExampleCleanDirty
Simple example to explain Adapton's dirtying + cleaning algorithms.

LazyFilter

LazyMap

LazyMergesort1

LazyMergesort2

LazyMergesort3

ListReverse

ListTree

ListTreeMax

ListTreeSum

Quickhull

RazDouble

RazMax

SimpEagerMap

UniformInsert

UniformPrepend

VecMax
Native Rust lab that finds the maximum random integer in an array.

Functions

all_labs
all_labs: This is the master list of all labs in the current Adapton Lab catalog (which consists of this module, but which wraps the Adapton crate's collections module).