Module main::catalog::hammer_s17_hw1

Hammer - CSCI 7000, Spring 2017

First homework assignment: #HW1

There are five functions below whose bodies consist of panic!("TODO"). Using the types listed in their declarations, implement these functions.

Test the behavior of your filter function, for instance:

cargo run -- -L hammer-s17-hw1-filter

Structs

Editor

LLEditor

RunFilter

RunJoin

RunReverse

RunSingletons

RunSplit

Enums

List
Compared with hw0, Cons cells only carry an element; Name and reference cell (Art) cases are separate from Cons.

Functions

list_filter
List filter:

list_join
List join:

list_reverse
List reverse:

list_singletons
List singletons:

list_split
List split: