Struct main::clusters::grid5000::Grid5000

pub struct Grid5000 {
    api_base_url: &'static str,
    deploy_url: &'static str,
    job_url_pretty: &'static str,
    job_url: &'static str,
    username: String,
    password: String,
    site: String,
    nb_nodes: String,
    walltime: String,
    ssh_key_path: String,
    reserved_node_address: String,
    uid: String,
}

Representation of a Grid5000 job reservation

Fields
api_base_url: &'static strdeploy_url: &'static strjob_url_pretty: &'static strjob_url: &'static strusername: Stringpassword: Stringsite: Stringnb_nodes: Stringwalltime: Stringssh_key_path: Stringreserved_node_address: Stringuid: String
Implementations

impl Grid5000

pub fn new(
username: String,
password: String,
site: String,
walltime: String
) -> Grid5000

pub fn new_random_site(
username: String,
password: String,
walltime: String
) -> Grid5000

Create a reservation on a random site that has available green energy

pub fn has_green_energy_available(self) -> bool

pub fn delete_reservation(&self, job_to_delete: String) -> Result<(), Error>

Delete reservation of node with uid = job_uid

fn reserve_node(&self) -> Result<JobSubmitResponse, Error>

Make a request to the Grid5000 to reserve a node

fn make_reservation(&mut self) -> String

Make a reservartion and return the adress of the reserved node

fn get_reservation(&self, job_uid: String) -> Result<JobSubmitResponse, Error>

Check state of reservation with uid = job_uid

fn deploy_env_on_node(
&self,
target_nodes: &Vec<String>,
environment: String,
ssh_key: String
) -> Result<DeployEnvResponse, Error>

Deploy provided environment to specified node

fn get_deployment(
&self,
deployment_uid: String
) -> Result<DeployEnvResponse, Error>

Check state of deployment with uid = deployment_uid

fn get_ssh_key(&self) -> Result<String, Box<dyn Error + 'static>>

Get the SSH key from provided file

fn get_sites_with_green_energy(&self) -> Vec<String>

Uses the OpenWeatherMap api to get the Grid5000 with available green energy

fn choose_random_site_with_green_energy(&self) -> String

Chooses a random grid5000 site with available green energy