Struct main::cpu::mmu::Flags

pub struct Flags {
    // some fields omitted
}

Methods

impl Flags

fn empty() -> Flags

Returns an empty set of flags.

fn all() -> Flags

Returns the set containing all flags.

fn bits(&self) -> uint

Returns the raw value of the flags currently stored.

fn from_bits(bits: uint) -> Option<Flags>

Convert from underlying bit representation, unless that representation contains bits that do not correspond to a flag.

fn from_bits_truncate(bits: uint) -> Flags

Convert from underlying bit representation, dropping any bits that do not correspond to flags.

fn is_empty(&self) -> bool

Returns true if no flags are currently stored.

fn is_all(&self) -> bool

Returns true if all flags are currently set.

fn intersects(&self, other: Flags) -> bool

Returns true if there are flags common to both self and other.

fn contains(&self, other: Flags) -> bool

Returns true all of the flags in other are contained within self.

fn insert(&mut self, other: Flags)

Inserts the specified flags in-place.

fn remove(&mut self, other: Flags)

Removes the specified flags in-place.

impl Flags

fn empty() -> Flags

Returns an empty set of flags.

fn all() -> Flags

Returns the set containing all flags.

fn bits(&self) -> uint

Returns the raw value of the flags currently stored.

fn from_bits(bits: uint) -> Option<Flags>

Convert from underlying bit representation, unless that representation contains bits that do not correspond to a flag.

fn from_bits_truncate(bits: uint) -> Flags

Convert from underlying bit representation, dropping any bits that do not correspond to flags.

fn is_empty(&self) -> bool

Returns true if no flags are currently stored.

fn is_all(&self) -> bool

Returns true if all flags are currently set.

fn intersects(&self, other: Flags) -> bool

Returns true if there are flags common to both self and other.

fn contains(&self, other: Flags) -> bool

Returns true all of the flags in other are contained within self.

fn insert(&mut self, other: Flags)

Inserts the specified flags in-place.

fn remove(&mut self, other: Flags)

Removes the specified flags in-place.