main::data_structure::city

Struct City

pub struct City {
    name: String,
    heuristic_value: f64,
    latitude: f64,
    longitude: f64,
}

Estrutura que representa uma cidade no grafo.

Atributos:

name: Nome da cidade.

heuristic_value: Valor heurístico da cidade, usado em algoritmos de busca.

visited: Indica se a cidade foi visitada durante a busca.

latitude: Latitude da cidade.

longitude: Longitude da cidade.

Fields
name: Stringheuristic_value: f64latitude: f64longitude: f64
Implementations

impl City

pub fn new( name: String, latitude: f64, longitude: f64, latitude_destiny: f64, longitude_destiny: f64, ) -> City

Construtor para criar uma nova cidade com nome, coordenadas e valores heurísticos.

pub fn get_heuristic_value(&self) -> f64

Retorna o valor heurístico da cidade.

pub fn get_latitude(&self) -> f64

Retorna a latitude da cidade.

pub fn get_longitude(&self) -> f64

Retorna a longitude da cidade.

pub fn set_heuristic_value(&mut self, heuristic_value: f64)

Define o valor heurístico da cidade.

pub fn set_latitude(&mut self, latitude: f64)

Define a latitude e longitude da cidade.

pub fn set_longitude(&mut self, longitude: f64)

Define a longitude da cidade.

pub fn get_name(&self) -> &str

Retorna o nome da cidade.

pub fn set_name(&mut self, name: String)

Define o nome da cidade.

pub fn to_string(&self) -> String

Retorna uma representação em string da cidade.