Struct main::jobs_buffer::jobs_buffer::JobsBuffer

pub struct JobsBuffer {
    jobs: Vec<BufferedJob>,
}

Keeps a list of jobs and submissions (which are jobs

Fields
jobs: Vec<BufferedJob>

The list of submissions and jobs (which are submissions sent to the photogrammetry service)

Implementations

impl JobsBuffer

pub fn new() -> JobsBuffer

Creates a JobsBuffer struct

pub fn add_job_or_submission(
&mut self,
job: BufferedJob
) -> Result<(), BufferError>

Adds a submission or a job to the buffer

pub fn remove_job(&mut self, id: &str) -> Result<(), BufferError>

Removes a job based on its id

pub fn get_job_by_id(&mut self, id: &str) -> Option<&mut BufferedJob>

Returns a job based on it's id

pub fn get_pending_submissions(&mut self) -> Option<Vec<&mut BufferedJob>>

Returns the list of jobs that have not been sent to be processed yet

pub fn submission_exists(&self, job: &BufferedJob) -> bool

Returns true if there's a submission in the buffer that has the same submission_id

pub fn job_exists(&self, job: &BufferedJob) -> bool

Returns true if there is a job in the buffer with the same id

pub fn has_buffered_jobs(&self) -> bool

Returns true if the buffer has jobs waiting to be processed

pub fn check_timeouts(&mut self)

Checks whether there are jobs that were created more than 24h ago.

If it finds any, the jobs will be set as a pending submission