Module main::kernel::mm::allocator

Mechanisms for the allocation of kernel objects.

Structs

Alloc

BuddyAlloc
The [buddy memory allocation[1]][1] system is implemented with the use of a binary tree. It can be augmented with the use of linked lists[[2]].

Traits

Allocator
The allocator interface. Based on an unfinished RFC.