Struct main::kernel::mm::allocator::BuddyAlloc

pub struct BuddyAlloc {
    pub order: uint,
    pub tree: Bitv,
}

The buddy memory allocation[1] system is implemented with the use of a binary tree. It can be augmented with the use of linked lists[2].

Buddy memory allocation - Wikipedia

52.206 Operating Systems. Heap memory allocation

Fields

order

tree

Methods

impl BuddyAlloc

fn new(order: uint, storage: Bitv) -> BuddyAlloc