Struct main::orchestrator::Orchestrator

pub struct Orchestrator {
    ticks_delay: u64,
    green_energy_timeout: u64,
    services_keeper: Arc<RwLock<ServicesKeeper>>,
    jobs_buffer: Arc<RwLock<JobsBuffer>>,
    clusters_manager: Arc<RwLock<ClustersManager>>,
    image_storage: Arc<ImageStorageService>,
    photogrammetry: Arc<PhotogrammetryService>,
    result_storage: Arc<ResultStorageService>,
}

The brain of the application. Its purpose is to orchestrate all the microservices while following energetic requirements

Fields
ticks_delay: u64

Delay in seconds between iterations of the orchestrator main loop (cf Orchestrator::start())

green_energy_timeout: u64

Delay in seconds before forcing jobs processing without waiting for green energy

services_keeper: Arc<RwLock<ServicesKeeper>>

Keeps information to access microservices (hostname, port, username, password)

jobs_buffer: Arc<RwLock<JobsBuffer>>

Keeps the list of ongoing submissions and jobs. Note: a job is a submission that has been sent to the photogrammetry service

clusters_manager: Arc<RwLock<ClustersManager>>

Keeps the list of clusters where the photogrammetry service can be deployed in

image_storage: Arc<ImageStorageService>

Client for the images storage service, which stores the images of end users submissions

photogrammetry: Arc<PhotogrammetryService>

Client for the photogrammetry microservice

result_storage: Arc<ResultStorageService>

Client for the results storage microservice, which stores the result of photogrammetry calculations

Implementations

impl Orchestrator

pub fn new(
ticks_delay: u64,
green_energy_timeout: u64,
services_keeper: Arc<RwLock<ServicesKeeper>>,
jobs_buffer: Arc<RwLock<JobsBuffer>>,
clusters_manager: Arc<RwLock<ClustersManager>>,
image_storage: Arc<ImageStorageService>,
photogrammetry: Arc<PhotogrammetryService>,
result_storage: Arc<ResultStorageService>
) -> Orchestrator

Constructs an Orchestrator struct

pub fn start(orchestrator: Arc<Orchestrator>)

Starts the orchestrator main loop

It uses an Arc as it must be able to call the Orchestrator::start_web_server() function

fn deploy_and_register_photogrammetry_service(
&self,
selected_cluster: &mut Box<dyn ClusterFeatures + Send + Sync>
)

Deploys the photogrammetry service and registers its access information in the services keeper

fn start_web_server(orchestrator: Arc<Orchestrator>)

Starts a web service that listens to the 7878 port to make the orchestrator abe to handle pings from other microservices

It uses an Arc to allow using the orchestrator in different threads, which is necessary to handle TCP connections

fn add_submissions_to_buffer(&self) -> Result<(), String>

Fetches the new submissions from the ImageStorageService and adds them to the JobsBuffer if it doesn't have them already

fn select_jobs_to_run<'a>(
&self,
jobs: &'a mut [&'a mut BufferedJob],
selected_cluster: &'a Box<dyn ClusterFeatures + Send + Sync>
) -> Option<Vec<&'a mut BufferedJob>>

Decides which jobs to run in the list of pending submissions based on available green energy produced for the selected cluster

fn run_jobs(&self, jobs: &mut [&mut BufferedJob]) -> Result<(), String>

Sends all the received jobs to the photogrammetry service

fn handle_tcp_connection(&self, stream: TcpStream) -> Result<(), String>

Handles basic tcp connections

fn photogrammetry_callback(&self, id: &str) -> Result<(), ServiceError>

Reacts to a ping from the photogrammetry microservice when a job is done :

Sends the result url to the result storage microservice

Sets the status of the submission to 'Done' in the image storage service

Removes the submission from the buffer