Struct main::rust::data::allocator::Allocator

pub struct Allocator<T: 'static + Alloc + Sync + Sized> {
    // some fields omitted
}

Simple mechanism for allocating objects from a pool.

Methods

impl<T: Alloc> Allocator<T>

unsafe fn new(pool: *mut T, refs: *mut usize, num: usize, alloc_err: &'static str, dealloc_err: &'static str) -> Allocator<T>

Constructs an allocator from a pool of objects, a place to store pointers to objects in the pool, and panic errors for allocate and deallocate.

Safety

This is unsafe because it assumes that pool and refs are at least num in length and both zero.

fn available(&self) -> usize

Returns the number of available objects in this allocator.

fn has_available(&self) -> bool

Returns the same value as available() > 0, but may be faster if the implementation allows it.

unsafe fn alloc_uninitialized(&mut self) -> Allocated<T>

Allocates an object from its allocator.

Panics

Panics if !has_available().

Safety

This is unsafe because it doesn't initialize the allocated object.

unsafe fn alloc_unitialized_option(&mut self) -> Option<Allocated<T>>

Opitonally allocates an object from its allocator.

Safety

This is unsafe because it doesn't initialize the allocated object.

fn alloc(&mut self) -> Allocated<T> where T: Init

Allocates an object from its allocator.

Panics

Panics if !has_available().

fn alloc_option(&mut self) -> Option<Allocated<T>> where T: Init

Optionally allocates an object from its allocator.