Trait main::rust::data::container::generic::Container

pub trait Container<'a, T: 'a> where Self: ContainerImpl<'a, T> {
    fn from_full(data: &'a mut [T]) -> Self { ... }
    unsafe fn from_existing(data: &'a mut [T], len: usize) -> Self { ... }
    unsafe fn from_empty(data: &'a mut [T]) -> Self { ... }
    unsafe fn from_full_ptr(data: *mut T, cap: usize) -> Self { ... }
    unsafe fn from_existing_ptr(data: *mut T, cap: usize, len: usize) -> Self { ... }
    unsafe fn from_empty_ptr(data: *mut T, cap: usize) -> Self { ... }
    fn len(&self) -> usize { ... }
    fn capacity(&self) -> usize { ... }
    fn is_empty(&self) -> bool { ... }
    fn is_full(&self) -> bool { ... }
    fn peek_mut(&mut self) -> Option<&mut T> { ... }
    fn peek_mut_or_panic(&mut self, err: &'static str) -> &mut T { ... }
    fn peek(&self) -> Option<&T> { ... }
    fn peek_or_panic(&self, err: &'static str) -> &T { ... }
    fn remove(&mut self) -> Option<T> { ... }
    fn remove_or_panic(&mut self, err: &'static str) -> T { ... }
    fn remove_and_zero(&mut self) -> Option<T> { ... }
    fn remove_and_zero_or_panic(&mut self, err: &'static str) -> T { ... }
    fn forget(&mut self, n: usize) -> usize { ... }
    fn forget_and_zero(&mut self, n: usize) -> usize { ... }
    fn add_or_panic(&mut self, value: T, err: &'static str) { ... }
    fn extend(&mut self, values: &mut [T]) -> usize { ... }
    fn extend_or_panic(&mut self, values: &mut [T], err: &'static str) { ... }
    fn clear(&mut self) { ... }
    fn clear_and_zero(&mut self) { ... }
    unsafe fn move_data(&mut self, data: &'a mut [T]) { ... }
    unsafe fn move_data_and_zero(&mut self, data: &'a mut [T]) { ... }
}

Generic container; automatically extends ContainerImpl.

Provided Methods

fn from_full(data: &'a mut [T]) -> Self

Constructs a container an existing array of data.

unsafe fn from_existing(data: &'a mut [T], len: usize) -> Self

Constructs a container from an existing array of data.

Panics

Panics if len > data.len().

Safety

This is unsafe because the elements of data[len..] are not dropped, even if len < data.len().

unsafe fn from_empty(data: &'a mut [T]) -> Self

Constructs a container from an empty array of data.

Safety

This is unsafe because the elements of data are not dropped.

unsafe fn from_full_ptr(data: *mut T, cap: usize) -> Self

Constructs a container from a pointer to an existing array of data.

Safety

This is unsafe because it assumes that data is cap long and initialized.

unsafe fn from_existing_ptr(data: *mut T, cap: usize, len: usize) -> Self

Constructs a container from a pointer to an existing array of data.

Safety

This is unsafe because it assumes that the array is uninitialised after the first len elements and can contain at least cap elements.

unsafe fn from_empty_ptr(data: *mut T, cap: usize) -> Self

Constructs a container from a pointer to an uninitialised array.

Safety

This is unsafe because it assumes that the array is uninitialised and can contain at least cap elements.

fn len(&self) -> usize

Returns the current number of elements in the container in O(1) time.

fn capacity(&self) -> usize

Returns the maximum number of elements that the container can hold in O(1) time.

fn is_empty(&self) -> bool

Returns the same value as len() > 0, but may be faster if the implementation allows it.

fn is_full(&self) -> bool

Returns the same value as len() == capacity(), but may be faster if the implementation allows it.

fn peek_mut(&mut self) -> Option<&mut T>

Returns a mutable reference to the next element in the container which would be removed by a call to remove().

fn peek_mut_or_panic(&mut self, err: &'static str) -> &mut T

Returns a mutable reference to the next element in the container which would be removed by a call to remove().

Panics

Panics with err if is_empty().

fn peek(&self) -> Option<&T>

Returns an immutable reference to the next element in the container which would be removed by a call to remove().

fn peek_or_panic(&self, err: &'static str) -> &T

Returns an immutable reference to the next element in the container which would be removed by a call to remove().

Panics

Panics with err if is_empty().

fn remove(&mut self) -> Option<T>

Removes a value from the container and returns the value, or None if the container is empty.

fn remove_or_panic(&mut self, err: &'static str) -> T

Removes a value from the container and returns the value.

Panics

Panics with err if is_empty().

fn remove_and_zero(&mut self) -> Option<T>

Removes a value from the container and returns the value, or None if the container is empty. Zeroes out the value's original location in memory.

fn remove_and_zero_or_panic(&mut self, err: &'static str) -> T

Removes a value from the container and returns the value. Zeroes out the value's original location in memory.

Panics

Panics with err if is_empty().

fn forget(&mut self, n: usize) -> usize

Removes multiple values from the container and returns how many values were removed.

fn forget_and_zero(&mut self, n: usize) -> usize

Removes multiple values from the container and returns how many values were removed. Zeroes out the values' original locations in memory.

fn add_or_panic(&mut self, value: T, err: &'static str)

Adds an element to the container.

Panics

Panics with err if is_full().

fn extend(&mut self, values: &mut [T]) -> usize

Adds multiple elements to the container and returns how many values were added. May be faster than multiple adds if the implementation allows it.

Safety

This is unsafe because it copies the original values in the array instead of moving them.

fn extend_or_panic(&mut self, values: &mut [T], err: &'static str)

Adds multiple elements to the container. May be faster than multiple adds if the implementation allows it.

Panics

Panics if the value was partially added to the container.

Safety

This is unsafe because it copies the original values in the array instead of moving them.

fn clear(&mut self)

Drops all of the elements in the container.

fn clear_and_zero(&mut self)

Drops all of the elements in the container. Zeroes out the container's buffer.

unsafe fn move_data(&mut self, data: &'a mut [T])

Copies the container's data to a new pointer. Guarantees that the data will be contiguous after the copy.

Panics

Panics if data.len() < self.len().

Safety

This is unsafe because there will be two mutable references to the same data, and because data[..self.len] is not dropped.

unsafe fn move_data_and_zero(&mut self, data: &'a mut [T])

Copies the container's data to a new pointer. Guarantees that the data will be contiguous after the copy.

Panics

Panics if data.len() < self.len().

Safety

This is unsafe because data[..self.len] is not dropped.