Module main::rust::data::container::prio_queue

Structs

PrioQueue
A fixed-size, greatest-priority-first queue.

Traits

Priority
Trait for determining priority.

Functions

prio_max
Compares and returns the maximum of two values, by priority.

prio_min
Compares and returns the minimum of two values, by priority.