Module main::structs

Reexports

use std::prelude::v1::*;

Structs

Key
A tuple of a state and an action (which is a string), which is used as a key in the q_values HashMap to find the value of taking a certain action while in a certain state

State
Describes the current state at which the system is in. At the moment it only allows for one i32 variable, but this can be extended easily.