Crate main

Reexports

extern crate std as std;

extern crate rand;

use std::prelude::v1::*;

use rand::Rng;

use structs::{State, Key};

use std::collections::HashMap;

use std::process::{Command, Stdio};

use std::io::prelude::*;

Modules

structs

Functions

decide
Depending on chance, outputs the actions with the highest reward, or decides on an action depending on the decision policy.

find_best
Looks through actions available and the player's current state, and returns a vector containing two actions that provide the top two rewards

main
Initializes parameters for Q-learning and runs the game infinitely.

run_game
Runs the game and updates the Q-Value table as it plays.