Crate main

This is a demo of hacspec for the HACS workshop on September 22.

Overview

There are 5 separate parts to this demo.

Demo hacspec for some primitive side by side with RFC pseudocode

Poly1305

Polynomial example

Demo the typechecker with some (self-introduced) bug

Demo F* translation with a more advanced bug found by F*

Demo F* spec side-by-side with existing HACL* spec, and a proof of equivalence for one function

Setup

The typechecker requires a Rust nightly compiler and additional Rust components.

rustup toolchain install nightly
rustup component add --toolchain nightly rustc-dev
rustup component add --toolchain nightly llvm-tools-preview

To run the typechecker and compiler on a spec check out the hacspec repository

git clone https://github.com/hacspec/hacspec.git

HACL* and F*

To typecheck the generated F* code F* and a checkout of HACL* is required.

git clone https://github.com/project-everest/hacl-star.git

Windows: note that this hasn't been tested on Windows.

Demo Specs

The demo specs are in separate crates:

ChaCha20 hacspec

Poly1305 hacspec

ChaCha20Poly1305 hacspec

Polynomial arithmetic

Running tests on the spec

Each spec has tests in the tests folder. They can be run with cargo test in the top level directory or each individual crate.

Typecheck specs

To ensure that a spec is compliant with the hacspec language the typechecker can be used. The typechecker also outputs F* code.

Go into hacspec/language run the compiler as follows on the examples in this repository (assuming the hacspec-hacs-sept-2020-demo is next to hacspec). You can point this to any hacspec implementation.

cargo +nightly run -- \
    -L $PWD/../hacspec-hacs-sept-2020-demo/target/debug/deps \
    --crate-type=lib --edition=2018 \
    --extern=hacspec_lib -Zno-codegen \
    $PWD/../../hacspec-hacs-sept-2020-demo/hacspec-examples/poly1305/poly1305.rs

To generate F* code from hacspec run

cargo +nightly run -- \
    -L $PWD/../hacspec-hacs-sept-2020-demo/target/debug/deps \
    --crate-type=lib --edition=2018 \
    --extern=hacspec_lib \
    -o $PWD/../../hacspec-hacs-sept-2020-demo/Hacspec.Poly1305.fst \
    $PWD/../../hacspec-hacs-sept-2020-demo/hacspec-examples/poly1305/poly1305.rs

Typecheck F* specs

The generated F* specs can be typechecked with F* using HACL* and the hacspec F* library.

Prerequisites:

fstar.exe is in your path

$HACL_HOME is set to a checkout HACL*

$HACSPEC_HOME is set to the checkout of hacspec

fstar.exe --lax \
    --include $HACSPEC_HOME/fstar \
    --include $HACL_HOME/lib \
    Hacspec.Poly1305.fst