Crate main

Magnes.ie project

Modules

clusters
Contains the clusters manager as well as all the clusters implementations

jobs_buffer
Contains the jobs buffer, which keeps track of the submissions and the jobs (which are submissions sent to the photogrammetry service)

meteo_service
Contains a client to make REST request to the OpenWeatherMap API to get weather data in different cities

orchestrator

services
Contains clients for all the microservices and the ServiceKeeper, which keeps track of where the microservices are

ssh_client
Contains a client to perform a SSH connection and SSH commands on remote host

Constants

VERBOSE
If set to true, displays logging in the standard output

Functions

add_clusters
Add clusters to the clusters manager

add_grid5000_cluster
Adds a g5k cluster to the clusters manager

log
Print a message to the standard output

log_error
Print an error to the standard error

main
Instantiates the various components and starts the Orchestrator