Crate main

password manager that syncs with Google Drive

So what I want to do is create a command line application which I can use as follows:

$ passman save for github.com user maheshbansod pass thegithubpassword
$ passman gen-pass [12]
$ passman save for <website> user <username> (pass <password>|genpass [<len>])
$ passman get for <website> user <username>
$ passman sync

Enums

Cli

Functions

main