Crate main

Buscaminas

Valentina Laura Correa - vcorrea@fi.uba.ar - 104415

Introducción

La presente entrega contiene las funcionalidades pedidas para el trabajo practico nº1 de la materia Taller de Programacion I - curso Deymonnaz.

Objetivo

El objetivo de este trabajo practico consta de simular la logica del juego Buscaminas implementandolo en Rust siguiendo los conceptos trabajados en clase.

Para acceder al repositorio donde fue desarrollado el mismo se puede visitar el siguiente enlace.

Ejecución

Para comenzar a utilizar el programa se deberá hacer uso del comando cargo run seguido de la ruta en donde se encuentra el archivo de entrada. En particular, los archivos de entrada estan dentro de /main/src/mapas, ruta en donde tambien se guardara el archivo de salida.

Ejemplo: cargo run mapas/mapa_input.txt

Otros comandos de interes:

cargo test

cargo fmt

cargo clippy

cargo doc –open

Modules

archivos_service 🔒

Modulo que se centra en las funcionalidades referentes a la interacción con los archivos.

buscaminas_service 🔒

Service principal, ya que es el que contiene la logica mas importante del buscaminas. Se encuentra modularizado en las distintas funcionalidades requeridas para lograr develar las minas.

calculadora_service 🔒

Este modulo tiene como objetivo centralizar las operaciones y calculos realizados en base a las filas y columnas de la matriz del Buscaminas.

interaccion_usuario 🔒

Modulo encargado de hacer mas amigable la interaccion con el programa, buscando mantener informado al usuario.

Constants

ERROR

Functions

main 🔒

Funcion principal que controla el flujo del programa relacionandose con diferentes modulos. Recibe por parametros la ruta del archivo contenedor del mapa. Escribe en un nuevo archivo su resolución. Se asume que el contenido del archivo llega en el formato correcto